Struct hyper_router::Router

pub struct Router {
    // some fields omitted
}

This is the one. The router.

Methods

impl Router

fn find_handler_with_defaults(&self, request: &Request) -> Handler

Finds handler for given Hyper request.

This method uses default error handlers. If the request does not match any route than default 404 handler is returned. If the request match some routes but http method does not match (used GET but routes are defined for POST) than default method not supported handler is returned.

fn find_handler(&self, request: &Request) -> HttpResult<Handler>

Finds handler for given Hyper request.

It returns handler if it's found or StatusCode for error. This method may return NotFound, MethodNotAllowed or NotImplemented status codes.

fn find_matching_routes(&self, request_path: &str) -> Vec<&Route>

Returns vector of Routes that match to given path.